feat: refactor import paths for components to use Containers directory structure
